+++ crates/blockifier/src/execution/syscalls/hint_processor.rs
@@ -426,7 +426,7 @@ impl HintProcessor for SyscallHintProcessor<'_> {
         &self,
         hint_code: &str,
         _ap_tracking_data: &ApTracking,
-        _reference_ids: &HashMap<String, usize>,
+        _reference_ids: &HashMap<String, u64>,
         _references: &[HintReference],
     ) -> Result<Box<dyn Any>, VirtualMachineError> {
         Ok(Box::new(self.hints[hint_code].clone()))
